Extract Listing of Blog/Projects into own lib function
